Move logging DuplicateFilter to utils, enable it only in __main__
